Dev nch

+++ src/backend/rest/roles.py
@@ -15,8 +15,9 @@
 
 # @dev : Nathan Cheval <nathan.cheval@outlook.fr>
 
+from flask_babel import gettext
 from flask import Blueprint, request, make_response, jsonify
-from src.backend.import_controllers import auth, roles
+from src.backend.import_controllers import auth, roles, privileges
 
 bp = Blueprint('roles', __name__, url_prefix='/ws/')
 
@@ -24,22 +25,19 @@ bp = Blueprint('roles', __name__, url_prefix='/ws/')
 @bp.route('roles/list', methods=['GET'])
 @auth.token_required
 def get_roles():
-    args = {
-        'select': ['*', 'count(*) OVER() as total'],
-        'offset': request.args['offset'] if 'offset' in request.args else 0,
-        'limit': request.args['limit'] if 'limit' in request.args else 'ALL'
-    }
-    if 'full' in request.args:
-        args['where'] = ['status NOT IN (%s)']
-        args['data'] = ['DEL']
+    if not privileges.has_privileges(request.environ['user_id'], ['settings', 'roles_list']):
+        return jsonify({'errors': gettext('UNAUTHORIZED_ROUTE'), 'message': '/roles/list'}), 403
 
-    _roles = roles.get_roles(args)
+    _roles = roles.get_roles(request.args)
     return make_response(jsonify(_roles[0])), _roles[1]
 
 
 @bp.route('roles/getById/<int:role_id>', methods=['GET'])
 @auth.token_required
 def get_role_by_id(role_id):
+    if not privileges.has_privileges(request.environ['user_id'], ['settings', 'update_role']):
+        return jsonify({'errors': gettext('UNAUTHORIZED_ROUTE'), 'message': f'/roles/getById/{role_id}'}), 403
+
     _role = roles.get_role_by_id(role_id)
     return make_response(jsonify(_role[0])), _role[1]
 
@@ -47,6 +45,9 @@ def get_role_by_id(role_id):
 @bp.route('roles/update/<int:role_id>', methods=['PUT'])
 @auth.token_required
 def update_role(role_id):
+    if not privileges.has_privileges(request.environ['user_id'], ['settings', 'update_role']):
+        return jsonify({'errors': gettext('UNAUTHORIZED_ROUTE'), 'message': f'/roles/update/{role_id}'}), 403
+
     data = request.json['args']
     res = roles.update_role(role_id, data)
     return make_response(jsonify(res[0])), res[1]
@@ -55,6 +56,9 @@ def update_role(role_id):
 @bp.route('roles/updatePrivilege/<int:role_id>', methods=['PUT'])
 @auth.token_required
 def update_privilege(role_id):
+    if not privileges.has_privileges(request.environ['user_id'], ['settings', 'update_role | add_role']):
+        return jsonify({'errors': gettext('UNAUTHORIZED_ROUTE'), 'message': f'/roles/updatePrivilege/{role_id}'}), 403
+
     data = request.json['privileges']
     res = roles.update_role_privilege(role_id, data)
     return make_response(jsonify(res[0])), res[1]
@@ -63,6 +67,9 @@ def update_privilege(role_id):
 @bp.route('roles/delete/<int:role_id>', methods=['DELETE'])
 @auth.token_required
 def delete_role(role_id):
+    if not privileges.has_privileges(request.environ['user_id'], ['settings', 'roles_list']):
+        return jsonify({'errors': gettext('UNAUTHORIZED_ROUTE'), 'message': f'/roles/delete/{role_id}'}), 403
+
     res = roles.delete_role(role_id)
     return make_response(jsonify(res[0])), res[1]
 
@@ -70,6 +77,9 @@ def delete_role(role_id):
 @bp.route('roles/disable/<int:role_id>', methods=['PUT'])
 @auth.token_required
 def disable_role(role_id):
+    if not privileges.has_privileges(request.environ['user_id'], ['settings', 'roles_list']):
+        return jsonify({'errors': gettext('UNAUTHORIZED_ROUTE'), 'message': f'/roles/disable/{role_id}'}), 403
+
     res = roles.disable_role(role_id)
     return make_response(jsonify(res[0])), res[1]
 
@@ -77,6 +87,9 @@ def disable_role(role_id):
 @bp.route('roles/enable/<int:role_id>', methods=['PUT'])
 @auth.token_required
 def enable_role(role_id):
+    if not privileges.has_privileges(request.environ['user_id'], ['settings', 'roles_list']):
+        return jsonify({'errors': gettext('UNAUTHORIZED_ROUTE'), 'message': f'/roles/enable/{role_id}'}), 403
+
     res = roles.enable_role(role_id)
     return make_response(jsonify(res[0])), res[1]
 
@@ -84,6 +97,9 @@ def enable_role(role_id):
 @bp.route('roles/create', methods=['POST'])
 @auth.token_required
 def create_role():
+    if not privileges.has_privileges(request.environ['user_id'], ['settings', 'add_role']):
+        return jsonify({'errors': gettext('UNAUTHORIZED_ROUTE'), 'message': '/roles/create'}), 403
+
     data = request.json['args']
     res = roles.create_role(data)
     return make_response(jsonify(res[0])), res[1]
